Рефакторинг ImagesListViewController. Перенос fateFormatter и showLikeError в ImageLustViewPresenter. Перенос форматирования даты и вызов алерта об ошибки  в презентер.
